requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cleo==0.8.1
clikit==0.6.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==3.2.1
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
filelock==3.0.12
html5lib==1.1
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
keyring==21.4.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
packaging==20.4
pastel==0.2.1
pexpect==4.8.0
pipenv==2020.8.13
pkginfo==1.6.1
poetry==1.1.4
poetry-core==1.0.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylev==1.3.0
pylint==2.6.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
shellingham==1.3.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.1
tomlkit==0.7.0
uritemplate==3.0.1
urllib3==1.25.11
validate-email==1.3
virtualenv==20.1.0
virtualenv-clone==0.5.4
webencodings==0.5.1
wrapt==1.12.1
